Clean-up test files.
