+++ test/Qowaiv.Validation.Fluent.UnitTests/Models/DateTimeModels.cs
@@ -1,22 +1,19 @@
-﻿using System;
+﻿namespace Qowaiv.Validation.Fluent.UnitTests.Models;
 
-namespace Qowaiv.Validation.Fluent.UnitTests.Models
+public class DateTimeModel
 {
-    public class DateTimeModel
-    {
-        public DateTime Prop { get; set; }
-    }
-    public class DateModel
-    {
-        public Date Prop { get; set; }
-    }
-    public class NullableDateModel
-    {
-        public Date? Prop { get; set; }
-    }
+    public DateTime Prop { get; set; }
+}
+public class DateModel
+{
+    public Date Prop { get; set; }
+}
+public class NullableDateModel
+{
+    public Date? Prop { get; set; }
+}
 
-    public class NullableDateTimeModel
-    {
-        public DateTime? Prop { get; set; }
-    }
+public class NullableDateTimeModel
+{
+    public DateTime? Prop { get; set; }
 }
