+++ test/Qowaiv.Validation.Fluent.UnitTests/Models/NoIPBasedEmailAddressModel.cs
@@ -1,13 +1,10 @@
-﻿using FluentValidation;
+﻿namespace Qowaiv.Validation.Fluent.UnitTests.Models;
 
-namespace Qowaiv.Validation.Fluent.UnitTests.Models
+public class NoIPBasedEmailAddressModel
 {
-    public class NoIPBasedEmailAddressModel
-    {
-        public EmailAddress Email { get; set; }
-    }
-    public class NoIPBasedEmailAddressModelValidator : ModelValidator<NoIPBasedEmailAddressModel>
-    {
-        public NoIPBasedEmailAddressModelValidator() => RuleFor(m => m.Email).NotIPBased();
-    }
+    public EmailAddress Email { get; set; }
+}
+public class NoIPBasedEmailAddressModelValidator : ModelValidator<NoIPBasedEmailAddressModel>
+{
+    public NoIPBasedEmailAddressModelValidator() => RuleFor(m => m.Email).NotIPBased();
 }
