+++ test/Qowaiv.Validation.Fluent.UnitTests/Models/RequiredModel.cs
@@ -1,19 +1,15 @@
-﻿using FluentValidation;
-using Qowaiv.Globalization;
+﻿namespace Qowaiv.Validation.Fluent.UnitTests.Models;
 
-namespace Qowaiv.Validation.Fluent.UnitTests.Models
+public class RequiredModel
 {
-    public class RequiredModel
+    public EmailAddress Email { get; set; }
+    public Country Country { get; set; }
+}
+public class RequiredModelValidator : ModelValidator<RequiredModel>
+{
+    public RequiredModelValidator()
     {
-        public EmailAddress Email { get; set; }
-        public Country Country { get; set; }
-    }
-    public class RequiredModelValidator : ModelValidator<RequiredModel>
-    {
-        public RequiredModelValidator()
-        {
-            RuleFor(m => m.Email).Required();
-            RuleFor(m => m.Country).Required(true);
-        }
+        RuleFor(m => m.Email).Required();
+        RuleFor(m => m.Country).Required(true);
     }
 }
