+++ test/Qowaiv.Validation.Fluent.UnitTests/Models/UnknownWithSeverityModel.cs
@@ -1,14 +1,11 @@
-﻿using FluentValidation;
+﻿namespace Qowaiv.Validation.Fluent.UnitTests.Models;
 
-namespace Qowaiv.Validation.Fluent.UnitTests.Models
+public class UnknownWithSeverityModel
 {
-    public class UnknownWithSeverityModel
-    {
-        public EmailAddress Email { get; set; }
-    }
-    public class UnknownWithSeverityModelValidator : ModelValidator<UnknownWithSeverityModel>
-    {
-        public UnknownWithSeverityModelValidator()
-            => RuleFor(m => m.Email).NotEmptyOrUnknown().WithSeverity(Severity.Warning);
-    }
+    public EmailAddress Email { get; set; }
+}
+public class UnknownWithSeverityModelValidator : ModelValidator<UnknownWithSeverityModel>
+{
+    public UnknownWithSeverityModelValidator()
+        => RuleFor(m => m.Email).NotEmptyOrUnknown().WithSeverity(Severity.Warning);
 }
