+++ test/Qowaiv.Validation.Fluent.UnitTests/Properties/GlobalUsings.cs
@@ -1,0 +1,9 @@
+﻿global using FluentAssertions;
+global using FluentValidation;
+global using NUnit.Framework;
+global using Qowaiv;
+global using Qowaiv.Globalization;
+global using Qowaiv.Validation.Fluent;
+global using Qowaiv.Validation.Fluent.UnitTests.Models;
+global using System;
+global using System.Globalization;
