+++ test/Qowaiv.Validation.Fluent.UnitTests/Validators/Unknown_specs.cs
@@ -1,0 +1,32 @@
+﻿namespace Validation.Unkown_specs;
+
+public class Set_value
+{
+    [Test]
+    public void Is_valid()
+        => new UnknownModel { Email = EmailAddress.Parse("info@qowaiv.corg") }
+        .Should().BeValidFor(new UnknownModelValidator());
+}
+
+public class Unknown
+{
+    [TestCase("'Email' must not be unknown.", "en-GB")]
+    [TestCase("'Email' mag niet onbekend zijn.", "nl-BE")]
+    public void Not_valid(string message, CultureInfo culture)
+    {
+        using (new CultureInfoScope(culture))
+        {
+            new UnknownModel { Email = EmailAddress.Unknown }
+                .Should().BeInvalidFor(new UnknownModelValidator())
+                .WithMessage(ValidationMessage.Error(message, "Email"));
+        }
+    }
+}
+
+public class Empty
+{
+    [Test]
+    public void Is_valid()
+        => new UnknownModel { Email = EmailAddress.Empty }
+        .Should().BeValidFor(new UnknownModelValidator());
+}
